storage/tests: extract batch generators
Factored out three generator classes (random_batches_generator,
key_limited_random_batch_generator, linear_int_kv_batch_generator) from
storage_test_fixture.h into a new batch_generators.h header file.

This is just nice to do, plus it will also be helpful when incrementally
migrating tests from Boost to Google Test.
